moved to proper directories, abstracted slider, fun volume UI
