+++ src/components/Slider/Slider.js
