+++ src/components/Video/VideoSettings.js
