+++ src/styles/components/_components.scss
@@ -10,6 +10,7 @@
 @import "rounded-box";
 @import "section";
 @import "separator";
+@import "slider";
 @import "sticky";
 @import "tile";
 @import "toast";
